backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.7.2
httpx==0.27.0
PyGithub==2.3.0
pydantic==2.7.4
pydantic-settings==2.3.4
cryptography==42.0.8
python-jose[cryptography]==3.3.0
